pyproject.toml
[project]
name = "srentusha"
version = "0.1.0"
description = "сринтюша"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.19.0",
    "httpx>=0.28.1",
    "motor>=3.7.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]
